fix issue displaying chat messages from players with specific name patterns
Fixes an issue with chat messages not being displayed when the chatting player has a name consisting of only color characters. This is due to an issue with the ignore system that appears to have been caused by a change to the Q_StripColor function introduced in b93817fe2b13dae387d2cfb50fccc9c4803e1171.